requirements.txt
# Enterprise Knowledge Base - Requirements
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.0
gradio==4.26.0
pydantic==2.6.0
pydantic-settings==2.1.0

# RAG & LLM
llama-index==0.11.0
llama-index-core==0.11.0
llama-index-embeddings-huggingface==0.2.0
llama-index-llms-ollama==0.2.0

# Vector Database
pymilvus==2.4.0
milvus-lite==2.4.0

# Graph Database
neo4j==5.18.0

# Embedding & Reranking
sentence-transformers==2.7.0
transformers==4.39.0
onnxruntime==1.17.0

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
unstructured==0.12.0
langdetect==1.0.9

# Database
aiosqlite==0.20.0

# Utilities
pyyaml==6.0.1
numpy==1.26.0
scikit-learn==1.4.0
aiofiles==23.2.1
python-multipart==0.0.9
httpx==0.27.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0

# Monitoring & Logging
prometheus-client==0.20.0

# Optional: FAISS backup
faiss-cpu==1.7.4
